Add message processors
Message processors are simple function receiving a `*Message` that
will do something with it. Two processors are provided `LogMessage` and
`printMessage`. The first one just logs it, and the second one prints
the output.

The `LogMessage` processor, defined in `slack` package, provides the
same functionality as the original code was, i.e. logging the message
event.

The `printMessage` processor, defined in main program, prints the
message event on stdout. This is provided just as an example.

Both are called as corresponding goroutines.