Fix non-breaking space parsing (#15)
* Fix bug on message parsing on missing job

When the job name was not in the list of jobs, the bot died. Now it
just reports back the error to the user, but keeps running.

* Fix non-breaking space parsing

When copying messages with mentions from Slack back to it, the copied
text uses a non-breaking space (Unicode \u00a0) right after the
mention. This is not treated like a regular space anymore, so the bot
can't parse the message correctly.

The `RemoveSpace` function iters over the string rebuilding the
string, as it's immutable, and when finds the non-breaking space
unicode, it replaces it with a regular space.