Fix parsing tests
The parsing tests would fail immediately if an error was expected, so
the value of the job return value or the arguments was never checked.

Fix the test logic to verify all return values in all cases and
introduce subtests for cleaner output when one of them fails. This
change caused one of the non-run tests to fail, because the parsing
fails and then no `job` value is returned. Fix the data for that test
and add a comment explaining the reson for the data in the test input.