Add context selector to getConfig
The function `getConfig` now requires one string parameter to select
the K8s' context to connect to. With an empty string, it will use the
default current context defined in kubecfg.

There is no way in `clientcmd.ClientConfig` to select a different
context to use once it was loaded. Even the client interfaces
provixded by the `client-go` package are missing such
feature.

Building a whole configuration object to support things like this,
looked like a harder path to take from this point. Particularly, as
the code is to be run on every command anyway.