Move k8s command registration to main
Part of the design is inflexible because slack gets injected
everywhere. This happens because there is a missing abstraction to
route the communication between the chat and the action clients.

By moving the responsibility of routing to the main function we make
this more visible and hopefully the proper abstraction will become
apparent.