Add text cleaning operation
Add a `CleanText()` helper to aggregate all text sanitizing operations
when we transform the message from slack into our internal
representation. This now includes trimming space at the ends of the
message.

We also get rid of the textual mention of the bot in the message, so
consumers don't need to as the Message structure captures if a mention
was made.