Move Slack initialization to main
The `Chat` object was responsible to initialize slack, which makes it
harder to test (no option to inject the client) and also grouped too
many responsibilities into the chat structure.

Moving the client initialization to main is a minor first step to make
the chat more generic and the setup easier to work with.