Pass test cases by value instead of pointer
The test cases are better passed by value than pointer, as with
pointers we can get into situations where the test fails to be
properly run due to an oversight.

This is an example of such case: In
d4009a22c71e3fbed0f35ff149b26b9e90fc77b4 we changed the behaviour of
`ReadMessage()` so mentions would no longer appear in the `text` field
of `Message` and we do not need consumers to do the cleanup
themselves. The tests still expect the mention to be present in the
output `text` field, but they do not fail, making the tests
unreliable.

This change would make the tests fail in the first testcase, and in
the last one (the one with a mention) so we comment-out the first and
adjust the last one.