Try to fix Github Actions
The golang version we determine is `^1.16` which effectively means the
latest availabel over 1.16. Therefore, we have the issue that some
changes in newer versions of Go have to be respected now.

The main failure of the actions is related to installing software with
`go get -u`, which is deprecated in favor of `go install` and
completely removed in 1.19 (or maybe earlier). So just replace the
former for the latter.

Another issue is that `go install` changes `go.mod` and `go.sum`, so
we need to do these installations before we checkout the code. Int eh
case of codeclimate, it also means we need to move the installation of
their agent tool to the step after checkout, or the path gets
overwritten.

When installing a tool with `go install`, if outside of a module you
must specify the version of the tool. We fix this by adding `@latest`
to the tools that had it missing to ensure we get the most recent
version of the tools installed.

Remaining failures are: a race condition in the test-suite, as we are
checking for race conditions in tests now, and staticcheck compains of
some deprecated packages from Go 1.16+ that become issues in newer
releases.